Resolving agents for a channel should ignore the channel's subset (and only rely on provided/ required capabilities)
In particular, agents do not have a subset, only channels have a subset with the purpose of using separate channels for different queries. If an agent is used to answer a query for a given channel should only rely on the capabilities provided by the agent, and the capabilities required by the channel.

Resolves #43